@@ -13,7 +13,7 @@ namespace Monai.Deploy.InformaticsGateway.DicomWeb.Client.API
         Dicom = 0,
         DicomJson = 1,
         DicomXml = 2,
-        OctetStreme = 3,
+        OctetStream = 3,
         ImageJpeg = 10,
         ImageGif = 11,
         ImagePng = 12,
@@ -34,7 +34,7 @@ namespace Monai.Deploy.InformaticsGateway.DicomWeb.Client.API
             { MimeType.Dicom, "application/dicom" },
             { MimeType.DicomJson, "application/dicom+json" },
             { MimeType.DicomXml, "application/dicom+xml" },
-            { MimeType.OctetStreme, "application/octet-stream" },
+            { MimeType.OctetStream, "application/octet-stream" },
             { MimeType.ImageJpeg, "image/jpeg" },
             { MimeType.ImageGif, "image/gif" },
             { MimeType.ImagePng, "image/png" },
