@@ -1,32 +1,18 @@
-﻿/*
- * Apache License, Version 2.0
- * Copyright 2019-2020 NVIDIA Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+﻿// SPDX-FileCopyrightText: © 2021-2022 MONAI Consortium
+// SPDX-FileCopyrightText: © 2019-2020 NVIDIA Corporation
+// SPDX-License-Identifier: Apache License 2.0
 
-using ConsoleAppFramework;
+using System;
+using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Monai.Deploy.InformaticsGateway.DicomWeb.Client;
 using Monai.Deploy.InformaticsGateway.DicomWeb.Client.API;
-using System;
-using System.Threading.Tasks;
 
 namespace Monai.Deploy.InformaticsGateway.DicomWeb.CLI
 {
-    internal class Program
+    internal static class Program
     {
         private static async Task Main(string[] args)
         {
@@ -34,10 +20,10 @@ namespace Monai.Deploy.InformaticsGateway.DicomWeb.CLI
             await Host.CreateDefaultBuilder()
                 .ConfigureLogging(logging =>
                 {
-                    logging.AddConsole(options =>
+                    logging.AddSimpleConsole(configure =>
                     {
-                        options.IncludeScopes = false;
-                        options.TimestampFormat = "hh:mm:ss ";
+                        configure.IncludeScopes = false;
+                        configure.TimestampFormat = "hh:mm:ss ";
                     });
 
                     // Configure MinimumLogLevel(CreaterDefaultBuilder's default is Warning).
