@@ -1,30 +1,15 @@
-﻿/*
- * Apache License, Version 2.0
- * Copyright 2019-2020 NVIDIA Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+﻿// SPDX-FileCopyrightText: © 2021-2022 MONAI Consortium
+// SPDX-FileCopyrightText: © 2019-2020 NVIDIA Corporation
+// SPDX-License-Identifier: Apache License 2.0
 
-using Ardalis.GuardClauses;
-using ConsoleAppFramework;
-using Dicom;
-using Microsoft.Extensions.Logging;
-using Monai.Deploy.InformaticsGateway.DicomWeb.Client.API;
-using Monai.Deploy.InformaticsGateway.DicomWeb.Client.Common;
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Threading.Tasks;
+using Ardalis.GuardClauses;
+using FellowOakDicom;
+using Microsoft.Extensions.Logging;
+using Monai.Deploy.InformaticsGateway.Client.Common;
+using Monai.Deploy.InformaticsGateway.DicomWeb.Client.API;
 
 namespace Monai.Deploy.InformaticsGateway.DicomWeb.Client.CLI
 {
@@ -43,25 +28,28 @@ namespace Monai.Deploy.InformaticsGateway.DicomWeb.Client.CLI
         [Command("study", "Retrieves instances within a study")]
         public async Task Study(
             [Option("r", "Uniform Resource Locator (URL) of the DICOMweb service")] string rootUrl,
-            [Option("u", "username for authentication with the DICOMweb service")] string username,
-            [Option("p", "password for authentication with the DICOMweb service")] string password,
             [Option("q", "query parameters. e.g. \"PatientID=Bob*,00080060=CT\"")] string query,
+            [Option("u", "username for authentication with the DICOMweb service")] string username = "",
+            [Option("p", "password for authentication with the DICOMweb service")] string password = "",
             [Option("d", "fields to include. e.g. [PatientName, 00080020].")] List<string> fieldsToInclude = null,
             [Option("f", "fuzzy matching")] bool fuzzyMatching = false,
             [Option("o", "output directory, default: current directory(.)", DefaultValue = ".")] string outputDir = "."
             )
         {
-            Uri rootUri;
-            ValidateOptions(rootUrl, out rootUri);
+            ValidateOptions(rootUrl, out var rootUri);
             ValidateOutputDirectory(ref outputDir);
 
             _dicomWebClient.ConfigureServiceUris(rootUri);
-            _dicomWebClient.ConfigureAuthentication(Utils.GenerateFromUsernamePassword(username, password));
+
+            if (!string.IsNullOrWhiteSpace(username) && !string.IsNullOrWhiteSpace(password))
+            {
+                _dicomWebClient.ConfigureAuthentication(Utils.GenerateFromUsernamePassword(username, password));
+            }
             _logger.LogInformation($"Querying studies...");
 
             var queryParameters = ParseQueryString(query);
 
-            await SaveJson(outputDir, _dicomWebClient.Qido.SearchForStudies<string>(queryParameters, fieldsToInclude, fuzzyMatching));
+            await SaveJson(outputDir, _dicomWebClient.Qido.SearchForStudies<string>(queryParameters, fieldsToInclude, fuzzyMatching)).ConfigureAwait(false);
         }
 
         private Dictionary<string, string> ParseQueryString(string query)
@@ -91,24 +79,8 @@ namespace Monai.Deploy.InformaticsGateway.DicomWeb.Client.CLI
 
             await foreach (var item in enumerable)
             {
-                await Utils.SaveJson(_logger, outputDir, item, DicomTag.StudyInstanceUID);
+                await Utils.SaveJson(_logger, outputDir, item, DicomTag.StudyInstanceUID).ConfigureAwait(false);
             }
-        }
-
-        private async Task SaveFiles(string outputDir, IAsyncEnumerable<DicomDataset> enumerable)
-        {
-            Guard.Against.NullOrWhiteSpace(outputDir, nameof(outputDir));
-            Guard.Against.Null(enumerable, nameof(enumerable));
-
-            var count = 0;
-            await foreach (var dataset in enumerable)
-            {
-                var file = new DicomFile(dataset);
-                var filename = file.Dataset.GetSingleValueOrDefault(DicomTag.StudyInstanceUID, $"unknown-study-{count}");
-                await Utils.SaveFiles(_logger, file, Path.Combine(outputDir, filename));
-                count++;
-            }
-            _logger.LogInformation($"Successfully saved {count} files.");
         }
 
         private void ValidateOutputDirectory(ref string outputDir)
